SeedNodeSnapshot knows the Dao
