+++ monitor/README.md
@@ -77,10 +77,10 @@ P2PNetworkLoad.run.torProxyPort=9061
 P2PNetworkLoad.run.historySize=500
 
 #P2PNetworkMessageSnapshot Metric
-P2PSeedNodeSnapshot.enabled=true
-P2PSeedNodeSnapshot.run.interval=24
-P2PSeedNodeSnapshot.run.hosts=3f3cu2yw7u457ztq.onion:8000, 723ljisnynbtdohi.onion:8000, fl3mmribyxgrv63c.onion:8000
-P2PSeedNodeSnapshot.run.torProxyPort=9062
+P2PSeedNodeWithDaoSnapshot.enabled=true
+P2PSeedNodeWithDaoSnapshot.run.interval=24
+P2PSeedNodeWithDaoSnapshot.run.hosts=3f3cu2yw7u457ztq.onion:8000, 723ljisnynbtdohi.onion:8000, fl3mmribyxgrv63c.onion:8000
+P2PSeedNodeWithDaoSnapshot.run.torProxyPort=9062
 
 #P2PMarketStats Metric
 P2PMarketStats.enabled=true
