+++ monitor/src/main/java/bisq/monitor/metric/P2PSeedNodeSnapshot.java
@@ -24,29 +24,14 @@ import bisq.monitor.OnionParser;
 import bisq.monitor.Reporter;
 import bisq.monitor.ThreadGate;
 
-import bisq.core.dao.monitoring.model.BlindVoteStateHash;
-import bisq.core.dao.monitoring.model.DaoStateHash;
-import bisq.core.dao.monitoring.model.ProposalStateHash;
-import bisq.core.dao.monitoring.network.messages.GetBlindVoteStateHashesRequest;
-import bisq.core.dao.monitoring.network.messages.GetBlindVoteStateHashesResponse;
-import bisq.core.dao.monitoring.network.messages.GetDaoStateHashesRequest;
-import bisq.core.dao.monitoring.network.messages.GetDaoStateHashesResponse;
-import bisq.core.dao.monitoring.network.messages.GetProposalStateHashesRequest;
-import bisq.core.dao.monitoring.network.messages.GetProposalStateHashesResponse;
 import bisq.core.proto.network.CoreNetworkProtoResolver;
 
 import bisq.network.p2p.CloseConnectionMessage;
 import bisq.network.p2p.NodeAddress;
-import bisq.network.p2p.network.CloseConnectionReason;
 import bisq.network.p2p.network.Connection;
 import bisq.network.p2p.network.MessageListener;
 import bisq.network.p2p.network.NetworkNode;
 import bisq.network.p2p.network.TorNetworkNode;
-import bisq.network.p2p.peers.getdata.messages.GetDataResponse;
-import bisq.network.p2p.peers.getdata.messages.PreliminaryGetDataRequest;
-import bisq.network.p2p.storage.payload.PersistableNetworkPayload;
-import bisq.network.p2p.storage.payload.ProtectedStorageEntry;
-import bisq.network.p2p.storage.payload.ProtectedStoragePayload;
 
 import bisq.common.proto.network.NetworkEnvelope;
 
@@ -54,25 +39,17 @@ import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.SettableFuture;
 
-import java.net.MalformedURLException;
-
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Random;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
 
 import lombok.extern.slf4j.Slf4j;
 
 import org.jetbrains.annotations.NotNull;
-
-import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * Contacts a list of hosts and asks them for all the data excluding persisted messages. The
@@ -84,13 +61,12 @@ import static com.google.common.base.Preconditions.checkNotNull;
  *
  */
 @Slf4j
-public class P2PSeedNodeSnapshot extends Metric implements MessageListener {
+abstract public class P2PSeedNodeSnapshot extends Metric implements MessageListener {
 
     private static final String HOSTS = "run.hosts";
     private static final String TOR_PROXY_PORT = "run.torProxyPort";
     Statistics statistics;
     final Map<NodeAddress, Statistics> bucketsPerHost = new ConcurrentHashMap<>();
-    private final Set<byte[]> hashes = new TreeSet<>(Arrays::compare);
     private final ThreadGate gate = new ThreadGate();
 
     /**
@@ -109,58 +85,24 @@ public class P2PSeedNodeSnapshot extends Metric implements MessageListener {
         void reset();
     }
 
-    /**
-     * Efficient way to count message occurrences.
-     */
-    private class Counter {
-        private int value = 0;
-
-        int value() {
-            return value;
-        }
-
-        void increment() {
-            value++;
-        }
-    }
-
-    /**
-     * Use a counter to do statistics.
-     */
-    private class MyStatistics  implements  Statistics<Counter> {
-
-        private final Map<String, Counter> buckets = new HashMap<>();
-
-        @Override
-        public Statistics create() {
-            return new MyStatistics();
-        }
-
-        @Override
-        public synchronized void log(Object message) {
-
-            // For logging different data types
-            String className = message.getClass().getSimpleName();
-
-            buckets.putIfAbsent(className, new Counter());
-            buckets.get(className).increment();
-        }
-
-        @Override
-        public Map<String, Counter> values() {
-            return buckets;
-        }
-
-        @Override
-        public synchronized void reset() {
-            buckets.clear();
-        }
-    }
-
     public P2PSeedNodeSnapshot(Reporter reporter) {
         super(reporter);
 
-        statistics = new MyStatistics();
+
+//        AppendOnlyDataStoreService appendOnlyDataStoreService,
+//        ProtectedDataStoreService protectedDataStoreService,
+//        ResourceDataStoreService resourceDataStoreService,
+//        Storage<SequenceNumberMap> sequenceNumberMapStorage) {
+//
+//        Set<byte[]> excludedKeys = dataStorage.getAppendOnlyDataStoreMap().keySet().stream()
+//                .map(e -> e.bytes)
+//                .collect(Collectors.toSet());
+//
+//        Set<byte[]> excludedKeysFromPersistedEntryMap = dataStorage.getProtectedDataStoreMap().keySet()
+//                .stream()
+//                .map(e -> e.bytes)
+//                .collect(Collectors.toSet());
+
     }
 
     @Override
@@ -175,21 +117,14 @@ public class P2PSeedNodeSnapshot extends Metric implements MessageListener {
         // clear our buckets
         bucketsPerHost.clear();
 
-        int height = 550000;
-
-        Random random = new Random();
-        send(networkNode, new PreliminaryGetDataRequest(random.nextInt(), hashes));
-
-        send(networkNode, new GetDaoStateHashesRequest(height, random.nextInt()));
-
-        send(networkNode, new GetProposalStateHashesRequest(height, random.nextInt()));
-
-        send(networkNode, new GetBlindVoteStateHashesRequest(height, random.nextInt()));
+        getRequests().forEach(getDataRequest -> send(networkNode, getDataRequest));
 
         report();
     }
 
-    private void send(NetworkNode networkNode, NetworkEnvelope message) {
+    abstract protected List<NetworkEnvelope> getRequests();
+
+    protected void send(NetworkNode networkNode, NetworkEnvelope message) {
 
         ArrayList<Thread> threadList = new ArrayList<>();
 
@@ -239,50 +174,7 @@ public class P2PSeedNodeSnapshot extends Metric implements MessageListener {
     /**
      * Report all the stuff. Uses the configured reporter directly.
      */
-    void report() {
-
-        // report
-        Map<String, String> report = new HashMap<>();
-        // - assemble histograms
-        bucketsPerHost.forEach((host, statistics) -> statistics.values().forEach((type, counter) -> report
-                .put(OnionParser.prettyPrint(host) + ".numberOfMessages." + type, String.valueOf(((Counter) counter).value()))));
-
-        // - assemble diffs
-        //   - transfer values
-        Map<String, Statistics> messagesPerHost = new HashMap<>();
-        bucketsPerHost.forEach((host, value) -> messagesPerHost.put(OnionParser.prettyPrint(host), value));
-
-        //   - pick reference seed node and its values
-        Optional<String> referenceHost = messagesPerHost.keySet().stream().sorted().findFirst();
-        Map<String, Counter> referenceValues = messagesPerHost.get(referenceHost.get()).values();
-
-        //   - calculate diffs
-        messagesPerHost.forEach(
-            (host, statistics) -> {
-                statistics.values().forEach((messageType, count) -> {
-                    try {
-                        report.put(OnionParser.prettyPrint(host) + ".relativeNumberOfMessages." + messageType,
-                                String.valueOf(((Counter) count).value() - referenceValues.get(messageType).value()));
-                    } catch (MalformedURLException ignore) {
-                        log.error("we should never got here");
-                    }
-                });
-                try {
-                    report.put(OnionParser.prettyPrint(host) + ".referenceHost", referenceHost.get());
-                } catch (MalformedURLException ignore) {
-                    log.error("we should never got here");
-                }
-            });
-
-        // cleanup for next run
-        bucketsPerHost.forEach((host, statistics) -> statistics.reset());
-
-        // when our hash cache exceeds a hard limit, we clear the cache and start anew
-        if (hashes.size() > 150000)
-            hashes.clear();
-
-        reporter.report(report, getName());
-    }
+    abstract void report();
 
     @Override
     public void onMessage(NetworkEnvelope networkEnvelope, Connection connection) {
@@ -296,79 +188,5 @@ public class P2PSeedNodeSnapshot extends Metric implements MessageListener {
         }
     }
 
-    private class Tuple {
-        private final long height;
-        private final byte[] hash;
-
-        Tuple(long height, byte[] hash) {
-            this.height = height;
-            this.hash = hash;
-        }
-    }
-
-    private Map<NodeAddress, Tuple> daoStateHashData = new ConcurrentHashMap<>();
-    private Map<NodeAddress, Tuple> daoProposalStateHashData = new ConcurrentHashMap<>();
-    private Map<NodeAddress, Tuple> daoBlindVoteStateHashData = new ConcurrentHashMap<>();
-
-    protected boolean treatMessage(NetworkEnvelope networkEnvelope, Connection connection) {
-        if (networkEnvelope instanceof GetDataResponse) {
-
-            Statistics result = this.statistics.create();
-
-            GetDataResponse dataResponse = (GetDataResponse) networkEnvelope;
-            final Set<ProtectedStorageEntry> dataSet = dataResponse.getDataSet();
-            dataSet.forEach(e -> {
-                final ProtectedStoragePayload protectedStoragePayload = e.getProtectedStoragePayload();
-                if (protectedStoragePayload == null) {
-                    log.warn("StoragePayload was null: {}", networkEnvelope.toString());
-                    return;
-                }
-
-                result.log(protectedStoragePayload);
-            });
-
-            Set<PersistableNetworkPayload> persistableNetworkPayloadSet = dataResponse
-                    .getPersistableNetworkPayloadSet();
-            if (persistableNetworkPayloadSet != null) {
-                persistableNetworkPayloadSet.forEach(persistableNetworkPayload -> {
-
-                    // memorize message hashes
-                    //Byte[] bytes = new Byte[persistableNetworkPayload.getHash().length];
-                    //Arrays.setAll(bytes, n -> persistableNetworkPayload.getHash()[n]);
-
-                    //hashes.add(bytes);
-
-                    hashes.add(persistableNetworkPayload.getHash());
-                });
-            }
-
-            checkNotNull(connection.getPeersNodeAddressProperty(),
-                    "although the property is nullable, we need it to not be null");
-            bucketsPerHost.put(connection.getPeersNodeAddressProperty().getValue(), result);
-            return true;
-        } else if (networkEnvelope instanceof GetDaoStateHashesResponse) {
-            // get last entry
-            DaoStateHash last = ((GetDaoStateHashesResponse) networkEnvelope).getStateHashes().get(((GetDaoStateHashesResponse) networkEnvelope).getStateHashes().size() - 1);
-
-            daoStateHashData.put(connection.getPeersNodeAddressProperty().getValue(), new Tuple(last.getHeight(), last.getHash()));
-
-            return true;
-        } else if (networkEnvelope instanceof GetProposalStateHashesResponse) {
-            // get last entry
-            ProposalStateHash last = ((GetProposalStateHashesResponse) networkEnvelope).getStateHashes().get(((GetProposalStateHashesResponse) networkEnvelope).getStateHashes().size() - 1);
-
-            daoProposalStateHashData.put(connection.getPeersNodeAddressProperty().getValue(), new Tuple(last.getHeight(), last.getHash()));
-
-            return true;
-        } else if(networkEnvelope instanceof GetBlindVoteStateHashesResponse) {
-            // get last entry
-            BlindVoteStateHash last = ((GetBlindVoteStateHashesResponse) networkEnvelope).getStateHashes().get(((GetBlindVoteStateHashesResponse) networkEnvelope).getStateHashes().size() - 1);
-
-            daoBlindVoteStateHashData.put(connection.getPeersNodeAddressProperty().getValue(), new Tuple(last.getHeight(), last.getHash()));
-            connection.shutDown(CloseConnectionReason.APP_SHUT_DOWN);
-
-            return true;
-        }
-        return false;
-    }
+    abstract protected boolean treatMessage(NetworkEnvelope networkEnvelope, Connection connection);
 }
